+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### New features
+
+* Make the thread factory of a server configurable. This makes it possible to start an embedded nrepl server in a google app engine java dev server.
+
 ## 0.8.3 (2020-10-25)
 
 ### Bugs fixed